Rename centerReducers to centerReducer
