+++ client/js/reducers/centerReducer.js
